split timeout checker
