+++ iotdb-core/confignode/src/main/java/org/apache/iotdb/confignode/manager/pipe/agent/task/PipeConfigNodeSubtask.java
@@ -189,10 +189,6 @@ public class PipeConfigNodeSubtask extends PipeAbstractConnectorSubtask {
       decreaseReferenceCountAndReleaseLastEvent(event, true);
 
     } catch (final PipeException e) {
-      if (Thread.interrupted()) {
-        LOGGER.warn(
-            "The thread was interrupted, and the confignode subtask being executed has timed out.");
-      }
       setLastExceptionEvent(event);
       if (!isClosed.get()) {
         throw e;
@@ -204,10 +200,6 @@ public class PipeConfigNodeSubtask extends PipeAbstractConnectorSubtask {
         clearReferenceCountAndReleaseLastEvent(event);
       }
     } catch (final Exception e) {
-      if (Thread.interrupted()) {
-        LOGGER.warn(
-            "The thread was interrupted, and the confignode subtask being executed has timed out.");
-      }
       setLastExceptionEvent(event);
       if (!isClosed.get()) {
         throw new PipeException(
