+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/agent/task/subtask/processor/PipeProcessorSubtask.java
@@ -313,7 +313,7 @@ public class PipeProcessorSubtask extends PipeReportableSubtask {
     // do nothing
   }
 
-  public boolean isTimeOut() {
+  public boolean isTimeout() {
     return false;
   }
 }
