+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/agent/task/subtask/processor/PipeProcessorSubtaskWorker.java
@@ -106,7 +106,7 @@ public class PipeProcessorSubtaskWorker extends WrappedRunnable {
         }
       }
 
-      if (subtask.isTimeOut()) {
+      if (subtask.isTimeout()) {
         // Clean up the timeout status and exit the current thread
         subtask.markTimeoutStatus(false);
         final String message =
